Complete fix for class instance property indexers
The previous fix, 14945bb23ef124da7527f687cf8f481324103ab2, only worked
for literals. This fixes the original issue - indexer accesses were not
working for non-public fields.

The fix is more of a hack - it removes the access level check, meaning
in PlayScript you can access protected/private members of a class by
using property indexers. The correct fix is to pass in the context to
the indexer and do the proper access level checks; but it is safe as a
quick fix since is more permissive than as3.