Put null/undefined key handling behind a compile flag
Since the special handling of null and undefined key values is an
as3 wart, I'm going to put it behind a `#if !DISABLE_NULL_KEYS` check,
with the as3 behavior defaulting to enabled. This also means null
checks which were removed were added back.

Also removed the FormatKeyForAs call from a few spots where it wasn't
necessary (due to being called earlier in the call stack).