Extended URLLoader loading mechanism to support configurable conversions of passed url to a remote url.
When the passed url is not of the http or https protocol, the load mechanism checks for the existence of the path in local file system, per the logic found in Player.TryResolveResourcePath, if no such path exists, a call to the newly provided method Player.ToRemotePath is made.  This method is dependent on function assigned to the configurable property Player.RemotePathConverter.  If no value is set, the original path is returned; otherwise the function delegate is invoked to generate and return a value.